Move store to redux folder
